Bug fixes, asprintf introduction, RELEASE option.
Several compilation warnings were removed.  The one that is left comes
from the utlist.h macros.  It seems that they really have one variable that
they declare and don't use, but for another variable the warning is not
really clear to me.

I introduced a cmake option RELEASE, which triggers -O3 optimisation level
when building eq.  By introducing this option, I found several meory_test
failures, which magically disappeared when I introduced asprintf.

Instead of manual counting the length of the resulting string with later
concatenation, I suggest to use asprintf, which does it for you.  It might
be a wee bit slower, but from the readability perspective it is much more
clean.

+++ tests/eq-memory.sh
@@ -20,5 +20,7 @@
 
 valgrind $1 $2
 filename="${2##*/}"
-rm ${filename%.*}.py
+if [ -f ${filename%.*}.py ]; then
+        rm ${filename%.*}.py
+fi
 exit 0
